refactor: перенести валидаторы для User в UserContolller
