+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -2,10 +2,8 @@ package ru.yandex.practicum.filmorate.storage.user;
 
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.FilmorateNotFoundException;
-import ru.yandex.practicum.filmorate.exception.FilmorateValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -16,23 +14,18 @@ public class InMemoryUserStorage implements UserStorage {
 
     private final HashMap<Long, User> users = new HashMap<>();
 
-    private static Long id = 1L;
+    private Long id = 1L;
 
     @Override
     public User create(User user) {
-        validateUserBirthday(user);
         Long id = getNextId();
         User newUser = user.withId(id);
-        if (user.getName() == null || user.getName().isEmpty()) {
-            newUser = newUser.withName(user.getLogin());
-        }
         users.put(id, newUser);
         return newUser;
     }
 
     @Override
     public User update(User user) {
-        validateUserBirthday(user);
         if (users.containsKey(user.getId())) {
             users.put(user.getId(), user);
             return user;
@@ -62,6 +55,7 @@ public class InMemoryUserStorage implements UserStorage {
 
         user.getFriends().add(friendId);
         friend.getFriends().add(userId);
+
         return user;
     }
 
@@ -85,24 +79,7 @@ public class InMemoryUserStorage implements UserStorage {
                 .collect(Collectors.toList());
     }
 
-    @Override
-    public List<User> getMutualFriends(Long userId, Long otherUserId) {
-        User user = get(userId);
-        User otherUser = get(otherUserId);
-
-        return user.getFriends().stream()
-                .filter(f -> otherUser.getFriends().contains(f))
-                .map(users::get)
-                .collect(Collectors.toList());
-
-    }
-
-    private static Long getNextId() {
+    private Long getNextId() {
         return id++;
-    }
-    private void validateUserBirthday(User user) {
-        if (user.getBirthday().isAfter(LocalDate.now())) {
-            throw new FilmorateValidationException("День рождения не может быть в будущем.");
-        }
     }
 }
