+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -19,6 +19,4 @@ public interface UserStorage {
     User removeFriend(Long userId, Long friendId);
 
     List<User> getUserFriends(Long userId);
-
-    List<User> getMutualFriends(Long userId, Long otherUserId);
 }
